Renamed Lobster to UUV and moved PybulletAPI to other package
